@@ -23,7 +23,7 @@
     "@types/electron-settings": "3.0.0",
     "@types/fs-extra": "5.0.1",
     "@types/jasmine": "2.8.8",
-    "@types/jquery": "3.3.1",
+    "@types/jquery": "3.3.29",
     "@types/jsonfile": "4.0.1",
     "@types/lodash": "4.14.121",
     "@types/node": "8.0.53",
@@ -74,7 +74,7 @@
     "ts-loader": "2.3.1",
     "ts-node": "7.0.1",
     "tslint": "5.9.1",
-    "typescript": "2.6.2",
+    "typescript": "2.8.4",
     "webpack": "3.12.0"
   },
   "pre-commit": [
