@@ -2,16 +2,13 @@ import { Injectable } from '@angular/core';
 import { CanActivate, Router } from '@angular/router';
 import { Keymap } from 'uhk-common';
 
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { of } from 'rxjs/observable/of';
 import { switchMap, tap } from 'rxjs/operators';
 
-import 'rxjs/add/operator/let';
-
 import { Store } from '@ngrx/store';
 
-import { AppState } from '../../../store';
-import { getKeymaps } from '../../../store/reducers/user-configuration';
+import { AppState, getDefaultKeymap } from '../../../store';
 
 @Injectable()
 export class KeymapEditGuard implements CanActivate {
@@ -20,10 +17,9 @@ export class KeymapEditGuard implements CanActivate {
 
     canActivate(): Observable<boolean> {
         return this.store
-            .let(getKeymaps())
+            .select(getDefaultKeymap)
             .pipe(
-                tap((keymaps: Keymap[]) => {
-                    const defaultKeymap = keymaps.find(keymap => keymap.isDefault);
+                tap((defaultKeymap: Keymap) => {
                     if (defaultKeymap) {
                         this.router.navigate(['/keymap', defaultKeymap.abbreviation]);
                     }
