@@ -1,199 +1,139 @@
 import { Action } from '@ngrx/store';
-import { Keymap, Macro } from 'uhk-common';
-import { UndoUserConfigData } from '../../models/undo-user-config-data';
+import { Keymap } from 'uhk-common';
+
 import { ChangeKeymapDescription } from '../../models/ChangeKeymapDescription';
 import { KeyActionRemap } from '../../models/key-action-remap';
+import { UndoUserConfigData } from '../../models/undo-user-config-data';
 
-export type KeymapAction =
-    KeymapActions.AddKeymapAction |
-    KeymapActions.DuplicateKeymapAction |
-    KeymapActions.EditKeymapNameAction |
-    KeymapActions.EditKeymapAbbreviationAction |
-    KeymapActions.LoadKeymapSuccessAction |
-    KeymapActions.SetDefaultAction |
-    KeymapActions.RemoveKeymapAction |
-    KeymapActions.SaveKeyAction |
-    KeymapActions.CheckMacroAction |
-    KeymapActions.EditDescriptionAction;
+export enum ActionTypes {
+    Add = '[Keymap] Add keymap',
+    Duplicate = '[Keymap] Duplicate keymap',
+    EditAbbr = '[Keymap] Edit keymap abbreviation',
+    EditName = '[Keymap] Edit keymap title',
+    SaveKey = '[Keymap] Save key action',
+    SetDefault = '[Keymap] Set default option',
+    Remove = '[Keymap] Remove keymap',
+    CheckMacro = '[Keymap] Check deleted macro',
+    LoadKeymaps = '[Keymap] Load keymaps',
+    LoadKeymapsSuccess = '[Keymap] Load keymaps success',
+    EditDescription = '[Keymap] Edit description',
+    UndoLastAction = '[Keymap] Undo last action',
+    Select = '[Keymap] Select keymap action'
+}
 
-export namespace KeymapActions {
-    export const ADD = '[Keymap] Add keymap';
+export class AddKeymapAction implements Action {
+    type = ActionTypes.Add;
 
-    export type AddKeymapAction = {
-        type: typeof ADD,
-        payload: Keymap
-    };
-
-    export const DUPLICATE = '[Keymap] Duplicate keymap';
-
-    export type DuplicateKeymapAction = {
-        type: typeof DUPLICATE,
-        payload: Keymap
-    };
-
-    export const EDIT_ABBR = '[Keymap] Edit keymap abbreviation';
-
-    export type EditKeymapAbbreviationAction = {
-        type: typeof EDIT_ABBR,
-        payload: {
-            abbr: string;
-            newAbbr: string;
-            name: string;
-        }
-    };
-
-    export const EDIT_NAME = '[Keymap] Edit keymap title';
-
-    export type EditKeymapNameAction = {
-        type: typeof EDIT_NAME,
-        payload: {
-            abbr: string;
-            name: string;
-        }
-    };
-
-    export const SAVE_KEY = '[Keymap] Save key action';
-
-    export type SaveKeyAction = {
-        type: typeof SAVE_KEY,
-        payload: {
-            keymap: Keymap;
-            layer: number;
-            module: number;
-            key: number;
-            keyAction: KeyActionRemap;
-        }
-    };
-
-    export const SET_DEFAULT = '[Keymap] Set default option';
-
-    export type SetDefaultAction = {
-        type: typeof SET_DEFAULT,
-        payload: string
-    };
-
-    export const REMOVE = '[Keymap] Remove keymap';
-
-    export type RemoveKeymapAction = {
-        type: typeof REMOVE,
-        payload: string
-    };
-
-    export const CHECK_MACRO = '[Keymap] Check deleted macro';
-
-    export type CheckMacroAction = {
-        type: typeof CHECK_MACRO,
-        payload: Macro
-    };
-
-    export const LOAD_KEYMAPS = '[Keymap] Load keymaps';
-    export const LOAD_KEYMAPS_SUCCESS = '[Keymap] Load keymaps success';
-
-    export type LoadKeymapSuccessAction = {
-        type: typeof LOAD_KEYMAPS_SUCCESS,
-        payload: Keymap[]
-    };
-
-    export const UNDO_LAST_ACTION = '[Keymap] Undo last action';
-
-    export type UndoLastAction = {
-        type: typeof UNDO_LAST_ACTION,
-        payload: UndoUserConfigData
-    };
-
-    export const EDIT_DESCRIPTION = '[Keymap] Edit description';
-
-    export class EditDescriptionAction {
-        type = EDIT_DESCRIPTION;
-
-        constructor(public payload: ChangeKeymapDescription) {
-
-        }
-    }
-
-    export function loadKeymaps(): Action {
-        return {
-            type: KeymapActions.LOAD_KEYMAPS
-        };
-    }
-
-    export function loadKeymapsSuccess(keymaps: Keymap[]): LoadKeymapSuccessAction {
-        return {
-            type: KeymapActions.LOAD_KEYMAPS_SUCCESS,
-            payload: keymaps
-        };
-    }
-
-    export function addKeymap(item: Keymap): AddKeymapAction {
-        return {
-            type: KeymapActions.ADD,
-            payload: item
-        };
-    }
-
-    export function setDefault(abbr: string): SetDefaultAction {
-        return {
-            type: KeymapActions.SET_DEFAULT,
-            payload: abbr
-        };
-    }
-
-    export function removeKeymap(abbr: string): RemoveKeymapAction {
-        return {
-            type: KeymapActions.REMOVE,
-            payload: abbr
-        };
-    }
-
-    export function duplicateKeymap(keymap: Keymap): DuplicateKeymapAction {
-        return {
-            type: KeymapActions.DUPLICATE,
-            payload: keymap
-        };
-    }
-
-    export function editKeymapName(abbr: string, name: string): EditKeymapNameAction {
-        return {
-            type: KeymapActions.EDIT_NAME,
-            payload: {
-                abbr: abbr,
-                name: name
-            }
-        };
-    }
-
-    export function editKeymapAbbr(name: string, abbr: string, newAbbr: string): EditKeymapAbbreviationAction {
-        return {
-            type: KeymapActions.EDIT_ABBR,
-            payload: {
-                name,
-                abbr,
-                newAbbr
-            }
-        };
-    }
-
-    export function saveKey(keymap: Keymap,
-                            layer: number,
-                            module: number,
-                            key: number,
-                            keyAction: KeyActionRemap): SaveKeyAction {
-        return {
-            type: KeymapActions.SAVE_KEY,
-            payload: {
-                keymap,
-                layer,
-                module,
-                key,
-                keyAction
-            }
-        };
-    }
-
-    export function checkMacro(macro: Macro): CheckMacroAction {
-        return {
-            type: KeymapActions.CHECK_MACRO,
-            payload: macro
-        };
+    constructor(public payload: Keymap) {
     }
 }
+
+export class DuplicateKeymapAction implements Action {
+    type = ActionTypes.Duplicate;
+
+    constructor(public payload: Keymap) {
+    }
+}
+
+export class EditKeymapAbbreviationAction implements Action {
+    type = ActionTypes.EditAbbr;
+
+    constructor(public payload: {
+        abbr: string;
+        newAbbr: string;
+        name: string;
+    }) {
+    }
+}
+
+export class EditKeymapNameAction implements Action {
+    type = ActionTypes.EditName;
+
+    constructor(public payload: {
+        abbr: string;
+        name: string;
+    }) {
+    }
+}
+
+export class SaveKeyAction implements Action {
+    type = ActionTypes.SaveKey;
+
+    constructor(public payload: {
+        keymap: Keymap;
+        layer: number;
+        module: number;
+        key: number;
+        keyAction: KeyActionRemap;
+    }) {
+    }
+}
+
+export class SetDefaultKeymapAction implements Action {
+    type = ActionTypes.SetDefault;
+
+    constructor(public payload: string) {
+    }
+}
+
+export class RemoveKeymapAction implements Action {
+    type = ActionTypes.Remove;
+
+    constructor(public payload: string) {
+    }
+}
+
+export class CheckMacroAction implements Action {
+    type = ActionTypes.CheckMacro;
+
+    constructor(public payload: number) {
+    }
+}
+
+export class LoadKeymapsAction implements Action {
+    type = ActionTypes.LoadKeymaps;
+}
+
+export class LoadKeymapSuccessAction implements Action {
+    type = ActionTypes.LoadKeymapsSuccess;
+
+    constructor(public payload: Keymap[]) {
+    }
+}
+
+export class EditDescriptionAction implements Action {
+    type = ActionTypes.EditDescription;
+
+    constructor(public payload: ChangeKeymapDescription) {
+    }
+}
+
+export class UndoLastAction implements Action {
+    type = ActionTypes.UndoLastAction;
+
+    constructor(public payload: UndoUserConfigData) {
+    }
+}
+
+export class SelectKeymapAction implements Action {
+    type = ActionTypes.Select;
+
+    constructor(public payload: string) {
+    }
+}
+
+export type Actions
+    = AddKeymapAction
+    | DuplicateKeymapAction
+    | EditKeymapAbbreviationAction
+    | EditKeymapNameAction
+    | SaveKeyAction
+    | SetDefaultKeymapAction
+    | RemoveKeymapAction
+    | CheckMacroAction
+    | LoadKeymapsAction
+    | LoadKeymapSuccessAction
+    | EditDescriptionAction
+    | UndoLastAction
+    | SelectKeymapAction
+    ;
